:recycle: パッケージ構成のリファクタリング
各ジャンルごとのパッケージ構成整理
