+++ app/api/v1/account/controllers/TenantController.scala
@@ -1,9 +1,14 @@
-package api.v1.account
+package api.v1.account.controllers
 
+import api.v1.account.models.Tenant
+import api.v1.account.resourcehandlers.TenantResourceHandler
+import api.v1.common.{RequestMarkerContext, SaifuDefaultActionBuilder}
 import javax.inject.Inject
 import play.api.Logger
+import play.api.http.FileMimeTypes
+import play.api.i18n.{Langs, MessagesApi}
 import play.api.libs.json.Json
-import play.api.mvc.{Action, AnyContent}
+import play.api.mvc._
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -14,7 +19,7 @@ class TenantController @Inject() (cc: TenantControllerComponents)(implicit
   private val logger = Logger(getClass)
 
   def index: Action[AnyContent] =
-    TenantAction.async { implicit request =>
+    saifuDefaultAction.async { implicit request =>
       logger.trace("index: ")
       tenantResourceHandler.find.map { tenants =>
         Ok(Json.toJson(tenants))
@@ -22,7 +27,7 @@ class TenantController @Inject() (cc: TenantControllerComponents)(implicit
     }
 
   def process: Action[AnyContent] =
-    TenantAction.async { implicit request =>
+    saifuDefaultAction.async { implicit request =>
       logger.trace("process: ")
       Tenant.createTenantInput.bindFromRequest.fold(
         failure => {
@@ -37,7 +42,7 @@ class TenantController @Inject() (cc: TenantControllerComponents)(implicit
     }
 
   def update: Action[AnyContent] =
-    TenantAction.async { implicit request =>
+    saifuDefaultAction.async { implicit request =>
       logger.trace("update: ")
       Tenant.updateTenantInput.bindFromRequest.fold(
         failure => {
@@ -52,7 +57,7 @@ class TenantController @Inject() (cc: TenantControllerComponents)(implicit
     }
 
   def delete: Action[AnyContent] =
-    TenantAction.async { implicit request =>
+    saifuDefaultAction.async { implicit request =>
       logger.trace("delete: ")
       Tenant.deleteTenantInput.bindFromRequest.fold(
         failure => {
@@ -67,12 +72,47 @@ class TenantController @Inject() (cc: TenantControllerComponents)(implicit
     }
 
   def show(id: String): Action[AnyContent] = {
-    TenantAction.async { implicit request =>
+    saifuDefaultAction.async { implicit request =>
       logger.trace(s"show: id = $id")
       tenantResourceHandler.lookup(id).map { tenant =>
         Ok(Json.toJson(tenant))
       }
     }
   }
+}
 
+/**
+  * Packages up the component dependencies for the tenant controller
+  * @param saifuDefaultActionBuilder
+  * @param tenantResourceHandler
+  * @param actionBuilder
+  * @param parsers
+  * @param messagesApi
+  * @param langs
+  * @param fileMimeTypes
+  * @param executionContext
+  */
+case class TenantControllerComponents @Inject() (
+    saifuDefaultActionBuilder: SaifuDefaultActionBuilder,
+    tenantResourceHandler: TenantResourceHandler,
+    actionBuilder: DefaultActionBuilder,
+    parsers: PlayBodyParsers,
+    messagesApi: MessagesApi,
+    langs: Langs,
+    fileMimeTypes: FileMimeTypes,
+    executionContext: ExecutionContext
+) extends ControllerComponents
+
+/**
+  * Base Controller for Tenant
+  * @param tcc TenantControllerComponents
+  */
+class TenantBaseController @Inject() (tcc: TenantControllerComponents)
+    extends BaseController
+    with RequestMarkerContext {
+  override protected def controllerComponents: ControllerComponents = tcc
+
+  def saifuDefaultAction: SaifuDefaultActionBuilder = tcc.saifuDefaultActionBuilder
+
+  def tenantResourceHandler: TenantResourceHandler = tcc.tenantResourceHandler
 }
